Move definition of MarkFinalResult to result module
